textAdapter 관련 공통 코드 -> TextModel.kt에 메서드 생성
